% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sdtm.R
\name{build_sdtm}
\alias{build_sdtm}
\title{Build SDTMplus data from main and supp
domains}
\usage{
build_sdtm(lib, dst, ext = c("any"))
}
\arguments{
\item{lib}{location for xpt/sas7bdat datasets}

\item{dst}{data set name}

\item{ext}{extension, could be xpt or sas7bdat, if any then automatically select one}
}
\value{
SDTMplus data
}
\description{
Build SDTMplus data from main and supp
domains
}
\examples{
\dontrun{
dmplus <- build_sdtm(dm)
}
}
