% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sas.R
\name{import_sas}
\alias{import_sas}
\title{Wrapper function to read SAS datasets in either sas7bdat or xpt
from a directory}
\usage{
import_sas(lib, dst, ext = c("any"))
}
\arguments{
\item{lib}{location for xpt/sas7bdat datasets}

\item{dst}{dataset name}

\item{ext}{extension, could be xpt or sas7bdat, if any then automatically select one}
}
\value{
sas dataset read by haven::read_sas
}
\description{
Wrapper function to read SAS datasets in either sas7bdat or xpt
from a directory
}
\examples{
\dontrun{
adsl <- import_sas(adam, adsl)
}
}
